requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
PyYAML~=6.0.2
httpx~=0.27.2
selenium~=4.26.1
pdfplumber~=0.11.4
Flask~=3.1.0
SQLAlchemy~=2.0.36
alembic~=1.14.0
python-dotenv~=1.0.1
aiogram~=3.15.0
Werkzeug~=3.1.3
PyPDF2~=3.0.1
psycopg2>=2.9
webdriver-manager
Flask-Cors~=5.0.0
celery~=5.3.6
redis~=5.0.1
flower~=2.0.1
PyJWT~=2.10.1
Jinja2~=3.1.4
boto3~=1.35.76